validation des tests pour la lecture et l'écriture dans un fichier

+++ nurl/FileManager.cs
@@ -8,6 +8,7 @@
  */
 using System;
 using System.IO;
+using System.Text;
 
 namespace nurl {
 	
@@ -23,11 +24,34 @@ namespace nurl {
 		}
 		
 		public FileManager (string path) {
-			this.fileStream = File.Create(path);
+			if (!fileExiste(path)) {
+				this.fileStream = File.Create(path);
+			} else {
+				this.fileStream = File.Open(path, FileMode.Open);
+			}
 		}
 		
 		public static Boolean fileExiste(string url) {
 			return File.Exists(url);
+		}
+		
+		public void write (string texte) {
+			if (this.fileStream != null) {
+				StreamWriter writer = new StreamWriter(this.fileStream, Encoding.UTF8);
+				writer.Write(texte);
+				writer.Flush();
+			}
+		}
+		
+		public string getContent () {
+			if (this.fileStream != null) {
+				this.fileStream.Position = 0;
+				StreamReader reader = new StreamReader(this.fileStream, Encoding.UTF8);
+				string fileContent = reader.ReadToEnd();
+				reader.Dispose();
+				return fileContent;
+			}
+			return null;
 		}
 	}
 }
